Add an option command for enabling/disabling various options for a given session
This can be used to, for example, enable and disable the autoplaylist while the bot is running. Most bool options are handled in a generic manner, while the autoplaylist (among perhaps others in future) have their own
statements to handle beginning to play the music and check other conditions. Closes #900, closes #1240, closes #1415. It doesn't exactly complete #60 yet, as it doesn't reload the config, but rather sets options for the
duration of the current bot session.