:sparkles: remove a specific song from somewhere in the queue (#568)
* added cmd_remove

* add remove_entry for cmd_remove

* fixed consistency in !help comment for cmd_remove

* added logic to catch rare index error

sometimes you can remove a song from the very end of the queue while the player loads a new song which causes the index to no longer exist. let me know if there is a better way to handle this.

* many fixes to cmd_remove

took all suggestions from review of the PR. the command is smaller and more efficient, and also follows the same embed style for emojis.