Add playnext command to bot.py
cmd_play is now _cmd_play and takes an additional flag. The flag is used
by the new cmd_play and cmd_playnext to determine which side of the
playlist songs get added to. Songs get added to the front of the queue
if the flag, "head", is true, otherwise they get added to the back of
the queue.