Small change
Changes the nature of handling the repeat command when no option is given. It toggles through playlist looping, song looping, and looping off. If when turned off the playlist is 2 or more it will say playlist is no longer repeating. If the playlist is only 1 song it will say song is no longer repeating.