Fix mentions in now playing
There's an issue where using mentions in a title doesn't convert and shows as <@numbers> instead. This ensures now playing mentions work correctly by putting it into a field instead of a title. I'm sure there's a better way to hack the name parameter for add_field but this seems work fine.